refactor: split commit generator into separate files
